requirements.txt
streamlit
google-generativeai
python-docx
requests